[cgroups2] Introduced macro to return with an error if a cgroup is root.
There are several cgroup controls that only exist for non-root cgroups.
E.g. "cpu.weight", "cpu.idle", "cpu-max"

We introduce a marco to return with an error from functions that should not be
called with the root cgroup. This will be used extensively when implementing
APIs for each of the cgroup v2 subsystems.